added openCL on float, spder fractal

+++ sources/complex.c
@@ -1,6 +1,6 @@
 #include "fractol.h"
 
-t_complex	init_complex(double re, double im)
+t_complex	init_complex(long double re, long double im)
 {
 	t_complex res;
